+++ scripts/ts-utils/initApis.ts
@@ -3,14 +3,14 @@ const { cryptoWaitReady } = require('@polkadot/util-crypto');
 //fetchApi is used to fetch data from the chain
 const wsFetchProvider = new WsProvider("wss://tee-staging.litentry.io:443");
 
-//syncApi is used to send transactions to the chain
-const wsSyncProvider = new WsProvider("ws://localhost:9944");
+//defaultAPI is used to send transactions to the chain
+const wsDefaultProvider = new WsProvider("ws://localhost:9944");
 
 export const initApi = async () => {
     const fetchApi = await ApiPromise.create({ provider: wsFetchProvider });
-    const syncApi = await ApiPromise.create({ provider: wsSyncProvider });
+    const defaultAPI = await ApiPromise.create({ provider: wsDefaultProvider });
     await cryptoWaitReady();
-    return { fetchApi, syncApi };
+    return { fetchApi, defaultAPI };
 }
